Split ssm_util into subpackages to enhance code clarity
