+++ probdiffeq/_ssm_util/blockdiag_models.py
@@ -1,0 +1,614 @@
+from probdiffeq._ssm_util import ssm_api, utilities
+from probdiffeq.backend import func, linalg, np, random, structs, tree
+from probdiffeq.backend.typing import Any, Array, Callable, Literal, Sequence, TypeVar
+from probdiffeq.util import cholesky_util
+
+__all__ = [
+    "BlockDiagConditional",
+    "BlockDiagLinearizationFactory",
+    "BlockDiagNormal",
+    "BlockDiagOdeTs0",
+    "BlockDiagOdeTs1",
+    "BlockDiagPriorFactory",
+    "BlockDiagTreeFlatten",
+]
+
+
+C = TypeVar("C", bound=Sequence)
+"""A type-variable for Sequence types.
+
+For example, this variable is used to type Taylor coefficients.
+"""
+
+
+class BlockDiagPriorFactory(ssm_api.AbstractPriorFactory):
+    """Implementation of block-diagonal prior constructors."""
+
+    def identity(self, template) -> ssm_api.LatentCond:
+        (d, ndim) = template.mean_flat.shape
+        m0 = np.zeros((d, ndim))
+        c0 = np.zeros((d, ndim, ndim))
+        noise = BlockDiagNormal(m0, c0, template.tree_flatten)
+        matrix = np.ones((d, 1, 1)) * np.eye(ndim, ndim)[None, ...]
+        return ssm_api.LatentCond.from_linop_and_noise(matrix, noise)
+
+    def wiener_integrated(
+        self,
+        tcoeffs_mean: C,
+        /,
+        *,
+        is_exact: C | bool,
+        inexact_eps: float,
+        diffuse_derivatives: int,
+        diffuse_eps: float,
+        base_scale: Array | None,
+    ):
+        tcoeffs_std = self._tcoeffs_standard_deviation(
+            tcoeffs_mean, is_exact=is_exact, inexact_eps=inexact_eps
+        )
+        return self.wiener_integrated_diffuse(
+            tcoeffs_mean,
+            tcoeffs_std,
+            diffuse_derivatives=diffuse_derivatives,
+            diffuse_eps=diffuse_eps,
+            base_scale=base_scale,
+        )
+
+    def wiener_integrated_diffuse(
+        self,
+        tcoeffs_mean: C,
+        tcoeffs_std: C,
+        /,
+        *,
+        diffuse_derivatives: int,
+        diffuse_eps: float,
+        base_scale: Array | None,
+    ):
+        if diffuse_derivatives > 0:
+            tcoeffs_mean, tcoeffs_std = self._add_diffuse_derivatives(
+                tcoeffs_mean,
+                tcoeffs_std,
+                diffuse_derivatives=diffuse_derivatives,
+                diffuse_eps=diffuse_eps,
+            )
+
+        # Construct the initial variable from the mean and std
+        init = BlockDiagNormal.from_mean_and_std(tcoeffs_mean, tcoeffs_std)
+
+        single_flat, single_unravel = tree.ravel_pytree(tcoeffs_mean[0])
+        coeff_like = np.ones_like(single_flat)
+        base_scale_expected = single_unravel(coeff_like)
+        if base_scale is None:
+            base_scale = base_scale_expected
+        else:
+            if base_scale.shape != base_scale_expected.shape:
+                msg = "The base-scale has the wrong shape."
+                msg += f" Expected: {base_scale_expected.shape}."
+                msg += f" Received: {base_scale.shape}."
+                raise ValueError(msg)
+
+        base_scale, _ = tree.ravel_pytree(base_scale)
+
+        num_derivatives = len(tcoeffs_mean) - 1
+        a, q_sqrtm = utilities.system_matrices_1d_iwp(num_derivatives)
+        q0 = np.zeros((num_derivatives + 1,))
+        precon_fun = utilities.preconditioner_taylor(num_derivatives=num_derivatives)
+
+        def discretise(dt, output_scale: Array | None = None):
+            p, p_inv = precon_fun(dt)
+            if output_scale is None:
+                output_scale = np.ones_like(base_scale)
+            else:
+                output_scale = np.asarray(output_scale)
+
+                if output_scale.shape != base_scale.shape:
+                    msg = "The output-scale has the wrong shape."
+                    msg += f" Expected: {base_scale.shape}."
+                    msg += f" Received: {output_scale.shape}."
+                    raise ValueError(msg)
+                output_scale, _ = tree.ravel_pytree(output_scale)
+
+            scale = output_scale * base_scale
+            # Flatten the scale into something compatible with the flattened SSM
+            (d,) = scale.shape
+
+            A_batch = np.ones((d, 1, 1)) * a[None, :, :]
+            mean = np.ones((d, 1)) * q0[None, :]
+            cholesky = scale[:, None, None] * q_sqrtm[None, :, :]
+            tree_flatten = BlockDiagTreeFlatten.from_example(tcoeffs_mean)
+            noise = BlockDiagNormal(mean, cholesky, tree_flatten)
+            p = np.ones((d, 1)) * p[None, :]
+            p_inv = np.ones((d, 1)) * p_inv[None, :]
+            return ssm_api.LatentCond(A_batch, noise, to_latent=p_inv, to_observed=p)
+
+        return init, discretise
+
+    def exponential(
+        self,
+        tcoeffs_mean: C,
+        /,
+        *,
+        vf_linear,
+        is_exact: C | bool,
+        inexact_eps: float,
+        diffuse_derivatives: int,
+        diffuse_eps: float,
+        base_scale: Array | None,
+    ):
+        del tcoeffs_mean
+        del vf_linear
+        del is_exact
+        del inexact_eps
+        del diffuse_derivatives
+        del diffuse_eps
+        del base_scale
+        msg = "Block-diagonal exponential priors have not been implemented (yet.)."
+        msg += " If you need them, reach out."
+        raise NotImplementedError(msg)
+
+    def exponential_diffuse(
+        self,
+        tcoeffs_mean: C,
+        tcoeffs_std: C,
+        /,
+        *,
+        vf_linear,
+        diffuse_derivatives: int,
+        diffuse_eps: float,
+        base_scale: Array | None,
+    ):
+        del tcoeffs_mean
+        del tcoeffs_std
+        del vf_linear
+        del diffuse_derivatives
+        del diffuse_eps
+        del base_scale
+        msg = "Block-diagonal exponential priors have not been implemented (yet.)."
+        msg += " If you need them, reach out."
+        raise NotImplementedError(msg)
+
+    def _tcoeffs_standard_deviation(self, tcoeffs_mean, /, *, is_exact, inexact_eps):
+
+        # Construct the initial std.
+        # If is_exact is a boolean, copy the pytree structure from the mean
+        # Otherwise, set the initial std element-wise.
+        if isinstance(is_exact, bool):
+            if is_exact:
+                tcoeffs_std = tree.tree_map(np.zeros_like, tcoeffs_mean)
+            else:
+
+                def eps_like(s):
+                    return inexact_eps * np.ones_like(s)
+
+                tcoeffs_std = tree.tree_map(eps_like, tcoeffs_mean)
+        else:
+
+            def std_init(s: Array) -> Array:
+                if s.dtype != np.dtype(bool):
+                    msg = "Boolean entries expected in `is_exact`."
+                    msg += f" Received: dtype={np.dtype(s)}"
+                    raise TypeError(msg)
+                return np.where(s, 0.0, inexact_eps)
+
+            tcoeffs_std = tree.tree_map(std_init, is_exact)
+
+        def shape_equal(A, B):
+            return tree.tree_map(lambda a, b: np.shape(a) == np.shape(b), A, B)
+
+        if not tree.tree_all(shape_equal(tcoeffs_mean, tcoeffs_std)):
+            msg = "Input 'is_exact' has the wrong PyTree structure."
+            msg += f" Expected: {tree.tree_map(np.shape, tcoeffs_mean)}."
+            msg += f" Received: {tree.tree_map(np.shape, is_exact)}."
+            raise ValueError(msg)
+
+        return tcoeffs_std
+
+    def _add_diffuse_derivatives(
+        self, tcoeffs_mean, tcoeffs_std, /, *, diffuse_derivatives, diffuse_eps
+    ):
+        # Always set the mean to zero (for now at least).
+        zeros = tree.tree_map(np.zeros_like, tcoeffs_mean[0])
+        tcoeffs_mean = [*tcoeffs_mean, *[zeros for _ in range(diffuse_derivatives)]]
+
+        unknowns = tree.tree_map(
+            lambda s: diffuse_eps * np.ones_like(s), tcoeffs_std[0]
+        )
+        tcoeffs_std = [*tcoeffs_std, *[unknowns for _ in range(diffuse_derivatives)]]
+        return tcoeffs_mean, tcoeffs_std
+
+    def to_derivative(self, i, std, template):
+
+        def select(a):
+            return np.asarray([a[i]])
+
+        (d, n) = template.mean_flat.shape
+        x = np.zeros((d, n))
+        linop = func.vmap(func.jacrev(select))(x)
+
+        u_like = tree.tree_map(np.zeros_like, template.mean[0])
+        noise = BlockDiagNormal.from_mean_and_std([u_like], [std])
+        return ssm_api.LatentCond.from_linop_and_noise(linop, noise)
+
+    def prototype_output_scale_calibrated(self, template):
+        single_flat, _ = tree.ravel_pytree(template.mean[0])
+
+        # TODO: technically, these should be pytrees according
+        # to the leaf structure, right?
+        return np.ones(single_flat.shape)
+
+
+class BlockDiagOdeTs0(ssm_api.AbstractOde):
+    """Construct a block-diagonal implementation of ODE-TS0 linearization."""
+
+    def __init__(self, vf, *, ode_order: int) -> None:
+        super().__init__(vf, ode_order=ode_order)
+
+    def init_linearization(self) -> None:
+        return None
+
+    def linearize(self, rv, state: None, *, damp: float, t):
+        del state
+        fx = self.vector_field(*(rv.mean[: self.ode_order]), t=t)
+        fx = tree.tree_map(lambda s: -s, fx)
+        bias = BlockDiagNormal.from_dirac([fx], damp=damp)
+
+        def a1(s):
+            return s[[self.ode_order], ...]
+
+        linop = func.vmap(func.jacrev(a1))(rv.mean_flat)
+
+        cond = ssm_api.LatentCond.from_linop_and_noise(linop, bias)
+        return cond, None
+
+
+class BlockDiagOdeTs1(ssm_api.AbstractOde):
+    """Construct a block-diagonal implementation of ODE-TS1 linearization."""
+
+    def __init__(self, vf, *, ode_order: int, jacobian: Any) -> None:
+        if ode_order > 1:
+            msg = "This linearization is not compatible with high-order ODEs as of yet."
+            raise ValueError(msg)
+        super().__init__(vf, ode_order=1)
+        self.jacobian = jacobian
+
+    def init_linearization(self):
+        return self.jacobian.init_jacobian_handler()
+
+    def linearize(self, rv, state, *, damp: float, t):
+        fun = func.partial(self.vector_field, t=t)
+
+        m0_tree = rv.mean[0]
+        rv0 = BlockDiagNormal.from_dirac([m0_tree], damp=0.0)
+
+        def a1(s):
+            return s[[1], ...]
+
+        linop = func.vmap(func.jacrev(a1))(rv.mean_flat)
+
+        def vf_flat(u):
+            u_tree = rv0.tree_flatten.unflatten_array(u[:, None])
+            fu_tree = fun(*u_tree)
+            return rv0.tree_flatten.flatten_tree([fu_tree]).reshape((-1,))
+
+        # Evaluate the linearisation
+        m0 = rv.mean_flat[:, 0]
+        fx, J_diag, state = self.jacobian.calculate_diagonal(vf_flat, m0, state)
+
+        E1 = func.jacrev(lambda s: s[0])(rv.mean_flat[0])
+        linop = linop - J_diag[:, None, None] * E1[None, None, :]
+
+        fx = rv.mean_flat[:, 1] - fx
+        fx = fx[..., None]
+        diff = func.vmap(lambda a, b: a @ b)(linop, rv.mean_flat)
+        fx = fx - diff
+        fx = rv0.tree_flatten.unflatten_array(fx)
+        bias = BlockDiagNormal.from_dirac([fx], damp=damp)
+        cond = ssm_api.LatentCond.from_linop_and_noise(linop, bias)
+        return cond, state
+
+
+class BlockDiagLinearizationFactory(ssm_api.AbstractLinearizationFactory):
+    """Construct a block-diagonal linearization-factory."""
+
+    def root(
+        self,
+        root,
+        *,
+        jacobian,
+        root_order: int | Literal["max"],
+        nlstsq: Callable | None,
+    ):
+        raise NotImplementedError
+
+    def ode_taylor_0th(self, vf, *, ode_order):
+        return BlockDiagOdeTs0(vf, ode_order=ode_order)
+
+    def ode_taylor_1st(self, vf, *, ode_order, jacobian):
+        return BlockDiagOdeTs1(vf, ode_order=ode_order, jacobian=jacobian)
+
+
+class BlockDiagConditional(ssm_api.AbstractConditional):
+    """Construct a block-diagonal implementation of manipulating conditionals."""
+
+    def apply_flat(self, x, cond, /):
+
+        def apply_unbatch(s, c):
+            s = c.to_latent * s
+            m_new = c.to_observed * (c.A @ s + c.noise.mean_flat)
+            c_new = c.to_observed[:, None] * c.noise.cholesky_flat
+            return BlockDiagNormal(m_new, c_new, cond.noise.tree_flatten)
+
+        return func.vmap(apply_unbatch)(x, cond)
+
+    def marginalise(self, rv, cond, /):
+        matrix, noise = cond.A, cond.noise
+        assert matrix.ndim == 3
+        mean = cond.to_latent * rv.mean_flat
+        cholesky = cond.to_latent[:, :, None] * rv.cholesky_flat
+
+        mean_marg = np.einsum("ijk,ik->ij", matrix, mean) + noise.mean_flat
+
+        chol1 = _transpose(matrix @ cholesky)
+        chol2 = _transpose(noise.cholesky_flat)
+        R_stack = (chol1, chol2)
+        cholesky = func.vmap(cholesky_util.sum_of_sqrtm_factors)(R_stack)
+
+        mean_new = cond.to_observed * mean_marg
+        cholesky_new = cond.to_observed[:, :, None] * _transpose(cholesky)
+        return BlockDiagNormal(mean_new, cholesky_new, cond.noise.tree_flatten)
+
+    def merge(self, cond1, cond2, /):
+        # Transform: latent (2) to latent (1)
+        T = cond1.to_latent * cond2.to_observed
+
+        # Linear operator
+        A1, A2 = cond1.A, T[:, :, None] * cond2.A
+        g = func.vmap(lambda a, b: a @ b)(A1, A2)
+
+        # Combined mean
+        m1, m2 = T * cond2.noise.mean_flat, cond1.noise.mean_flat
+        xi = func.vmap(lambda a, b, c: a @ b + c)(A1, m1, m2)
+
+        # Cholesky factor of combined covariance
+        C1, C2 = cond1.noise.cholesky_flat, T[:, :, None] * cond2.noise.cholesky_flat
+        R1 = _transpose(func.vmap(lambda a, b: a @ b)(A1, C2))
+        R2 = _transpose(C1)
+        Xi = func.vmap(cholesky_util.sum_of_sqrtm_factors)((R1, R2))
+        Xi = _transpose(Xi)
+
+        # Gather and return
+        noise = BlockDiagNormal(xi, Xi, cond1.noise.tree_flatten)
+        return ssm_api.LatentCond(
+            g, noise, to_latent=cond2.to_latent, to_observed=cond1.to_observed
+        )
+
+    def revert(self, rv, cond, /, *, solve_triu):
+        # Pull RV into latent space
+        mean = cond.to_latent * rv.mean_flat
+        cholesky = cond.to_latent[:, :, None] * rv.cholesky_flat
+
+        # QR decomposition
+        rv_chol_upper = _transpose(cholesky)
+        noise_chol_upper = _transpose(cond.noise.cholesky_flat)
+        A_rv_chol_upper = _transpose(cond.A @ cholesky)
+
+        revert_conditional = func.partial(
+            cholesky_util.revert_conditional, solve_triu=solve_triu
+        )
+        revert_vmap = func.vmap(revert_conditional)
+        r_obs, (r_cor, gain) = revert_vmap(
+            A_rv_chol_upper, rv_chol_upper, noise_chol_upper
+        )
+        cholesky_obs = np.transpose(r_obs, axes=(0, 2, 1))
+        cholesky_cor = np.transpose(r_cor, axes=(0, 2, 1))
+
+        # New backward conditional
+        mean_observed = (cond.A @ mean[..., None])[..., 0] + cond.noise.mean_flat
+        mean_corrected = mean - (gain @ (mean_observed[..., None]))[..., 0]
+        corrected = BlockDiagNormal(mean_corrected, cholesky_cor, rv.tree_flatten)
+        bwd = ssm_api.LatentCond(
+            gain,
+            corrected,
+            to_latent=1 / cond.to_observed,
+            to_observed=1 / cond.to_latent,
+        )
+
+        # Gather observed RV
+        mean_observed = cond.to_observed * mean_observed
+        cholesky_observed = cond.to_observed[:, :, None] * cholesky_obs
+        observed = BlockDiagNormal(
+            mean_observed, cholesky_observed, cond.noise.tree_flatten
+        )
+        return observed, bwd
+
+    def preconditioner_apply(self, cond, /):
+        A = cond.to_observed[:, :, None] * cond.A * cond.to_latent[:, None, :]
+        mean = cond.to_observed * cond.noise.mean_flat
+        cholesky = cond.to_observed[:, :, None] * cond.noise.cholesky_flat
+        noise = BlockDiagNormal(mean, cholesky, cond.noise.tree_flatten)
+        to_observed = np.ones_like(cond.to_observed)
+        to_latent = np.ones_like(cond.to_latent)
+        return ssm_api.LatentCond(
+            A, noise, to_observed=to_observed, to_latent=to_latent
+        )
+
+
+def _transpose(matrix):
+    return np.transpose(matrix, axes=(0, 2, 1))
+
+
+@structs.dataclass
+class BlockDiagTreeFlatten(ssm_api.AbstractTreeFlatten):
+    """Flattening information for block-diagonal random variables."""
+
+    # The treedef of the target
+    treedef: Any
+
+    # A map that unravels each leaf. Note that this is not the same
+    # as unravelling each Taylor coefficient, because the Taylor coefficients
+    # themselves can be pytrees, whereas the leaves are always arrays.
+    # The below function exclusively reshapes arrays
+    unravel_leaf: Any
+
+    def flatten_tree(self, x):
+        leaves = tree.tree_leaves_depth_one(x)
+        leaves_flat = [tree.ravel_pytree(s)[0] for s in leaves]
+        return np.stack(leaves_flat).T
+
+    def unflatten_array(self, x):
+        x_tree = tree.tree_unflatten(self.treedef, [*x.T])
+        return tree.tree_map(self.unravel_leaf, x_tree)
+
+    @classmethod
+    def from_example(cls, x):
+        leaves, treedef = tree.tree_flatten_depth_one(x)
+        _, unravel_leaf = tree.ravel_pytree(leaves[0])
+        return cls(treedef, unravel_leaf)
+
+
+class BlockDiagNormal(ssm_api.AbstractTreeNormal[BlockDiagTreeFlatten]):
+    """Construct a block-diagonal normal distribution.
+
+    This assumes that the pytree is of the form [M_1, ..., M_{num_coeffs}],
+    where each M_i is a pytree of the same structure.
+
+    Shapes:
+    - mean: (d, num_coeffs)
+    - cholesky: (d, num_coeffs, num_coeffs)
+
+    where d is the number of elements in each M.
+
+    """
+
+    @classmethod
+    def from_dirac(cls, mean, *, damp):
+        utilities.verify_taylor_coefficient_pytree(mean)
+
+        std = tree.tree_map(lambda s: damp * np.ones_like(s), mean)
+        return BlockDiagNormal.from_mean_and_std(mean, std)
+
+    @classmethod
+    def from_mean_and_std(cls, mean, std):
+        utilities.verify_taylor_coefficient_pytree(mean)
+        utilities.verify_taylor_coefficient_pytree(std)
+
+        tree_flatten = BlockDiagTreeFlatten.from_example(mean)
+
+        loc_flat = tree_flatten.flatten_tree(mean)
+        scale_flat = tree_flatten.flatten_tree(std)
+        num_coeffs = len(mean)
+
+        # Promote std into covariance matrix and apply damping
+        num_coeffs = len(mean)
+        d = np.ones((num_coeffs,))
+
+        cholesky = linalg.diagonal_matrix(d)
+        cholesky_flat = scale_flat[..., None] * cholesky[None, ...]
+        return cls(loc_flat, cholesky_flat, tree_flatten)
+
+    @property
+    def mean(self):
+        return self._mean_batched()
+
+    def _mean_batched(self):
+        if self.mean_flat.ndim > 2:
+            return func.vmap(BlockDiagNormal._mean_batched)(self)
+
+        return self.tree_flatten.unflatten_array(self.mean_flat)
+
+    @property
+    def std(self):
+        return self._std_batched()
+
+    def _std_batched(self):
+        if self.mean_flat.ndim > 2:
+            return func.vmap(BlockDiagNormal._std_batched)(self)
+
+        # TODO: use QR decomp instead of einsum to avoid sqrts
+        diag = np.einsum("ijk,ikj->ij", self.cholesky_flat, self.cholesky_flat)
+        std = np.sqrt(diag)
+        return self.tree_flatten.unflatten_array(std)
+
+    def residual_white_rms_tree(self, u, /):
+        # todo: add sth like an "axis" argument to make it more obvious
+        # to a user that this one here is vector-valued?
+
+        # assumes rv.chol = (d,1,1)
+        # return array of norms! See calibration
+        u_latent = self.tree_flatten.flatten_tree(u)
+        return self.residual_white_rms_flat(u_latent)
+
+    def residual_white_rms_flat(self, u_latent, /):
+        mean = np.reshape(self.mean_flat - u_latent, (-1,))
+        cholesky = np.reshape(self.cholesky_flat, (-1,))
+        return mean / cholesky / np.sqrt(mean.size)
+
+    def rescale_cholesky(self, factor, /):
+        cholesky = factor[..., None, None] * self.cholesky_flat
+        return BlockDiagNormal(self.mean_flat, cholesky, self.tree_flatten)
+
+    def logpdf_tree(self, u, /):
+        u_latent = self.tree_flatten.flatten_tree(u)
+        return self.logpdf_flat(u_latent)
+
+    def logpdf_flat(self, u_latent, /):
+        return np.sum(func.vmap(BlockDiagNormal.logpdf_scalar_flat)(self, u_latent))
+
+    def logpdf_scalar_flat(self, u):
+        cholesky = linalg.qr_r(self.cholesky_flat.T).T
+
+        dx = u - self.mean_flat
+        w = linalg.solve_triu(cholesky.T, dx, trans="T")
+
+        maha_term = linalg.vector_dot(w, w)
+
+        diagonal = linalg.diagonal_along_axis(cholesky, axis1=-1, axis2=-2)
+        slogdet = np.sum(np.log(np.abs(diagonal)))
+        logdet_term = 2.0 * slogdet
+        return -0.5 * (logdet_term + maha_term + u.size * np.log(np.pi() * 2))
+
+    def to_multivariate_normal(self):
+        mean = np.reshape(self.mean_flat.T, (-1,), order="F")
+        cov = np.block_diag(self._cov_dense())
+        return mean, cov
+
+    def _cov_dense(self):
+        if self.cholesky_flat.ndim > 2:
+            return func.vmap(BlockDiagNormal._cov_dense)(self)
+        return self.cholesky_flat @ self.cholesky_flat.T
+
+    def sample_tree(self, key):
+        if self.cholesky_flat.ndim > 3:
+            d, *_ = self.cholesky_flat.shape
+            keys = random.split(key, num=d)
+            return func.vmap(BlockDiagNormal.sample_tree)(self, keys)
+
+        sample_latent = self.sample_flat(key)
+        return self.tree_flatten.unflatten_array(sample_latent)
+
+    def sample_flat(self, key):
+        if self.cholesky_flat.ndim > 3:
+            d, *_ = self.cholesky_flat.shape
+            keys = random.split(key, num=d)
+            return func.vmap(BlockDiagNormal.sample_flat)(self, keys)
+
+        d, _n, n = self.cholesky_flat.shape
+        base = random.normal(key, shape=(d, n))
+        return self.mean_flat + np.einsum("ijk,ij->ik", self.cholesky_flat, base)
+
+    @staticmethod
+    def register_pytree_node() -> None:
+        def flatten(normal):
+            children = normal.mean_flat, normal.cholesky_flat
+            aux = (normal.tree_flatten,)
+            return children, aux
+
+        def unflatten(aux, children):
+            (tree_flatten,) = aux
+            mean, cholesky = children
+            return BlockDiagNormal(mean, cholesky, tree_flatten)
+
+        tree.register_pytree_node(BlockDiagNormal, flatten, unflatten)
+
+
+BlockDiagNormal.register_pytree_node()
